requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug>=3.0.0
email-validator==2.1.1
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==8.0.0
